レビュー指摘修正（自身が出品した商品情報を削除条件追加）

+++ app/controllers/items_controller.rb
@@ -39,8 +39,11 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    item.destroy
-    redirect_to root_path
+
+    if item.user_id == current_user.id
+      item.destroy
+      redirect_to root_path
+    end
   end
 
   private
